Extracted ArmyRepositorySqlite
